Removes `uuid4()` from within parametrized test cases.
Vscode test discovery caches the test case names including parametrized values.
This makes having `uuid4()` calls in the parametrized test cases an issue as vscode shows test case failures when it cannot find test cases that it has previously resolved.
As the change doesn't affect the utility of the test, it is better to fix the case for vscode users so we don't get the same issue reported in the future.