Only add new routes to the route_map
This avoids re-adding routes which are already in the route map.

Before, every route would be re-added on each call to
`construct_route_map` (which is called on each call to `register`).
`register` was called once per route handler in `Router.__init__`, once
for the openapi handler (if configured), and once per static files
configuration, as well as whenever a new route is registered outside the
`__init__` call. Instead, this keeps a set of routes which have already
been added to the route_map, and avoids re-adding them when calling
`construct_route_map`

This also avoids a bug when `register` is called after a static path has
been registered (when more than one static path is registered, or
`.register` is called to add an additional handler), where
`ImproperlyConfiguredException` with message:
`Cannot have configured routes below a static path` would be incorrectly
thrown when trying to add the static path to the route map a second
time.