Added Route Handlers To Path Mapper (#454)
Added route handlers to path mapper

Sometimes, decorators may require to resolve the endpoint when given an instance of route handler. Decorators have direct access to the route handler instance so by using this dictionary, it will be able to resolve its endpoint.

This is similar to `Flask.view_functions` in flask.