Implements `on_app_config` hooks.
Registered hooks must receive and return an instance of `AppConfig` object.

The `AppConfig` object is first instantiated from parameters passed to `Starlite.__init__()`, and then passed to each of the registered hook handlers.

The resultant `AppConfig` instance is used to instantiate the application.

Removes the use of `pydantic.validate_arguments()` on `Starlite.__init__()` as the instantiation of the `AppConfig` object does the same thing.

Test to ensure that the parameters of the `Starlite` constructor are matched by properties on `AppConfig`.

Test that ensures all properties on the `AppConfig` object are accessed within the `Starlite` constructor.