Convert optional collection types as early as possible.
Any parameter that we did something like `thing or []` with inside `Starlite.__init__()` and `Router.__init__()`, I've moved the `thing or []` bit to the `AppConfig` instantiation point.

This makes it nicer to use the `AppConfig` object externally, otherwise need to to the `if None` dance for every collection property on access.