OpenAPI (redoc/swagger/spotlight elements) usage in environments that cant access CDN's (#562)
* added option to use google fonts with redoc or not
added options to set css and js files for redoc, swagger and spotlight.

all for use in isolated environments. for example governments

* added option to use google fonts with redoc or not
added options to set css and js files for redoc, swagger and spotlight.

all for use in isolated environments. for example governments

added tests.

* added option to use google fonts with redoc or not
added options to set css and js files for redoc, swagger and spotlight.

all for use in isolated environments. for example governments

added tests. and updated docs to reflects this change

* added option to use google fonts with redoc or not
added options to set css and js files for redoc, swagger and spotlight.

all for use in isolated environments. for example governments

added tests. and updated docs to reflects this change

* fixed some typos in docs i added.

* updated the usage docs.

* fixed suggestions

* added tets for default CDN urls composing

* fixed a test i did not do correcttly

Co-authored-by: niels van hooij <niels.vanhooij@kpn.com>