Handling of compressed "body" in logging middleware.
If a compression middleware is configured, upon compression of a
response we add a flag to connection scope. Logging middleware checks
that flag to determine how logging of "body" should be handled.

Adds `exclude_compressed_body` to `LoggingMiddlewareConfig`.

If response body is compressed and `exclude_compressed_body` is `False`
we do not decode the body. If `True`, body is excluded from log.

Introduces a namespace for storing internal state in connection scope,
and utils `get_starlite_scope_state` and `set_starlite_scope_state`.