Handling of compressed "body" in logging middleware. (#822)
* Handling of compressed "body" in logging middleware.

If a compression middleware is configured, upon compression of a
response we add a flag to connection scope. Logging middleware checks
that flag to determine how logging of "body" should be handled.

Adds `exclude_compressed_body` to `LoggingMiddlewareConfig`.

If response body is compressed and `exclude_compressed_body` is `False`
we do not decode the body. If `True`, body is excluded from log.

Introduces a namespace for storing internal state in connection scope,
and utils `get_starlite_scope_state` and `set_starlite_scope_state`.

* Update starlite/utils/scope.py

* Update starlite/utils/scope.py

* Update reference docs.

* Update prose docs.

* Adds tests for {get,set}_starlite_scope_state utils.

* Appease the sonarcloud overlords.

* Change behavior of `get_starlite_scope_state()`.

If called with a default value, behaves like `dict.set_default()`.

If called without a default value, behaves like `dict.get()`.

* Rename STARLITE constant to SCOPE_STATE_NAMESPACE.

* Invert naming of config flag.

* Fix wording.