Fix #849 - Make LoggingMiddleware handle request bodies correctly (#860)
* Fix #849.

This was caused by trying to access a not yet available `body` of a request, and solved by passing the `receive` callable to the request extractor, allowing to be called should be body be requested. The rest of the stack should not be affected by this, since the received body is stored in the `scope`.