Use Starlite prod config with new mike deploy params
In `mike deploy`, I've added `--no-redirect` which makes `latest` resolve to a static URL path, instead of redirecting to the "latest" version.  This helps improve SEO by ensuring Google can serve a static URL that always shows the newest docs content.  All old versions will canonical to the `/latest/*` URL.

Also modified the title of the latest version to include "(latest)".  So in the docs front end, the version will print as "1.48 (latest)".