Ensure http requests with invalid payloads return 400.
At the point of parsing the `data` kwarg, this PR looks for a deserialization error
and raises a client error http response.

Adds a new exception, `SerializationException` which wraps the `msgspec` operations.
I think this better encapsulates the 3rd-party code.

Adds a new HTTP exception `ClientException` and makes other `4xx` errors a subclass.

Makes `ServiceUnavailableException` a subclass of `InternalServerException`.

Closes #1047