A mypy pre-commit hook for each sqla version.
Adds a 2nd mypy hook, aliased `mypy-sqla-1x`, that runs mypy with stubs
and plugin.

The existing `mypy` hook, runs checks without either the stubs or the
mypy plugin.

Requires an additional mypy config file so that we can specify the
plugin for the 1x env.

Also needs an alternate mypy cache for 1x as the two environments would
conflict with each other.