Optional `route_handlers` argument and accept any sequence of handlers.
Handlers don't _have_ to be provided by constructor. They can be provided by
plugin and added after instantiation of the application object. Also, the
pattern of passing an empty collection to the constructor is used a lot in
our tests.

Handlers don't have to be provided by a list to the constructor. Sequence
conveys that they are registered in the order that they are passed in,
and we convert to list when passing to the AppConfig. This allows 3rd party
plugins to modify the handlers if they need. Of course, 3rd party plugins
should be very explicit about what they do to the route handlers collection.