De-referencing forward referenced annotations for signature models. (#1263)
* De-referencing forward referenced annotations for signature models.

We currently rely on pydantic to resolve forward references for signature
models. However, in order to continue our migration away from pydantic we
should do this ourselves.

This change is a pre-cursor to add checks for improper use of reserved kwargs,
such as 'state'. In order to check the type that is annotated against the
reserved kwargs in function signatures, we need to have resolved any forward
references to their types first.

* Apply suggestions from code review

* Remove unnecessary names.

* Fixes unnecessary re-evaluation of fn type annotations.

* Fix annotation not in namespace.

* Remove type:ignore