Repository Enhancements (#1269)
* fix: adds an additional UUID ID mixin.  Generic repo type for one/many updates with the same method

* De-referencing forward referenced annotations for signature models. (#1263)

* De-referencing forward referenced annotations for signature models.

We currently rely on pydantic to resolve forward references for signature
models. However, in order to continue our migration away from pydantic we
should do this ourselves.

This change is a pre-cursor to add checks for improper use of reserved kwargs,
such as 'state'. In order to check the type that is annotated against the
reserved kwargs in function signatures, we need to have resolved any forward
references to their types first.

* Apply suggestions from code review

* Remove unnecessary names.

* Fixes unnecessary re-evaluation of fn type annotations.

* Fix annotation not in namespace.

* Remove type:ignore

* Cleanup `starlite/utils` (#1261)

* move utils/exceptions.py > middleware/exceptions/middleware.py

* move utils/extractors.py > middleware/logging/extractors.py

* update pre-commit

* move utils/csrf.py > middleware/csrf.py

* move utils/serialization.py > serialization.py

* remove utils/cli.py

* utils/model.py > utils/pydantic.py

* move utils/file.py > file_system.py

* move utils/pagination.py > pagination.py

* move middleware/logging/extractors.py > data_extractors.py

* move docs/examples/datastructures/pagination.py > docs/examples/pagination.py

* Check the type annotation of the `state` reserved kwarg. (#1264)

* De-referencing forward referenced annotations for signature models.

We currently rely on pydantic to resolve forward references for signature
models. However, in order to continue our migration away from pydantic we
should do this ourselves.

This change is a pre-cursor to add checks for improper use of reserved kwargs,
such as 'state'. In order to check the type that is annotated against the
reserved kwargs in function signatures, we need to have resolved any forward
references to their types first.

* Check the type annotation of the `state` reserved kwarg.

* Docs: Fix autorun examples (#1267)

fix autorun examples

* Xfail suspected fsspec windows issue (#1268)

xfail suspected fsspec windows issue

* feat: adds additional functions to repo pattern.

* fix: filters

* fix: typing adjustments adds test cases for insert many

* fix: linting errors

* fix: linting updates.  remove use of match for compatibility\

---------

Co-authored-by: Cody Fincher <cody@gluent.com>
Co-authored-by: Peter Schutt <peter@topsport.com.au>
Co-authored-by: Janek Nouvertné <25355197+provinzkraut@users.noreply.github.com>