Implement #1299: Accept header parsing and content negotiation
Implement a property to access the parsed value of the 'Accept' header
from the request.

The returned instance allows iteration in the preferred order of media types.
It also provides the 'best_match()' method, that can be used to implement
content negotiation by finding the best matching media type for a list
of provided types.