RFC: Remove `InitPluginProtocol` in favor of `on_app_init` hook.
We have two methods for automated app configuration, `on_app_init` hook, and `InitPluginProtocol`.
This PR attempts to simplify this by refactoring plugins to utilise the `on_app_init` hook, and
therefore removing the `InitPluginProtocol`.

In this PR, passing a plugin instance to `on_app_init` both adds the plugin to `AppConfig.plugins`
and performs any application configuration, such as that which is done by the SQLAlchemy plugin.

The `on_app_init` approach for pre-baked app config is desirable due to it interacting with the
`AppConfig` object. Plugin development is an area that we should encourage the community to
participate in, and encouraging that work to interact with the `AppConfig` _before_ the majority
of application configuration has taken place is more desirable than passing through an application
object to the plugin at some arbitrary point inside the app's own `__init__()` method.

Additionally, the app configuration is something that may be desired independent of a
serialization plugin. So it makes sense for that functionality to be defined separate from the
serialization plugin, and leveraged by it. For example, if a user doesn't wish to leverage the
serialization plugin functionality, they can use `on_app_init=[SQLAlchemyInit()]`. However if they
want the serialization behavior _and_ the app initialization, should that be
`on_app_init=[SQLAlchemyPlugin()]` or `plugins=[SQLAlchemyPlugin()]`, or
`on_app_init=[SQLAlchemyInit()], plugins=[SQLAlchemyPlugin()]`?

The `plugins` kwarg to app remains, so there is still two ways for a plugin to be registered,
e.g., by either passing `on_app_init=[plugin_instance]` or `plugins=[plugin_instance]`. This isn't
ideal, and definitely a downside of this approach.

Discussion points:
- is having plugin instances register themselves via `on_app_init` intuitive enough?
- does app configuration _have_ to be associated with serialization plugin behavior?
- an alternative would be to redefine `InitPluginProtocol.on_app_init()` to receive the app config
object and call these immediately after `on_app_init` is called.
- other??