Modify the `InitPluginProtocol` to interact with `AppConfig`.
This PR modifies `InitPluginProtocol.on_app_init()` to receive and return the `AppConfig` instance,
as opposed to the application instance itself.

The `Starlite.__init__()` method is modified to call the `on_app_init()` method of all registered
`InitPluginProtocol` plugins immediately after calling the `on_app_init` hooks.

This PR supersedes RFC PR #1381 which examined the effects of removing `InitPluginProtocol` in favor
of `on_app_init` hooks. From discussion in that PR it was resolved that there is merit in having
both `on_app_init` hooks and `InitPluginProtocol`, even if they essentially do the same thing, due
to improved ergonomics of the application interface that this affords.