Fix inconsistent parsed parameter annotations.
Params with `Annotated` types in handler signatures are parsed inconsistently,
depending on whether the annotation is stringized.

If the type annotation is a string, it is passed to `get_type_annotations()` which
unwraps any `Annotated` types. If the annotation is not a string, it remains as-is,
with `Annotated` in place.

This PR ensures that we parse `Annotated` types the same, irrespective of whether the
user has `__future__.annotations` imported, or not, by setting `include_extras=True`
on the call to `get_type_annotations()`.

We could consider explicitly unwrapping the annotated types and storing any metadata
on the `ParsedSignatureParameter` in a future PR.