RFC: Attempt at a `StarliteEncodableType` type alias. (#1425)
* RFC: Attempt at a `StarliteEncodableType` type alias.

`AbstractDTOInterface.to_encodable_type()` needs a type for its return value.

Alternative is to just type it `Any`, I don't know if we are tying a rod for
our backs trying to strong type this stuff, but then again, maybe it's ok.

* Add `collections.abc.Collection`.

This is to support `AbstractDTOInterface.to_encodable_type()` which may return
either a scalar or collection of types for encoding.

* Add to `starlite.types` namespace.

* Fix import for `TypeAlias`