RFC: A parsed signature datastructure.
This datastructure is intended to be the primary source of parameter and annotation
data for handler and dependency callables. Processing is only concerned with
rudimentary preparation of annotations that should _always_ occur before using
them, e.g., extracting annotations from `get_type_hints()`, extracting inner types
and metadata from `Annotated`, etc.

This is intended to be different from signature modelling which has aims that
conflict with having an absolute source of truth of the signature parameter and
return value types. For example, code such as this:

```py
...
elif parameter.should_skip_validation:
    attribute = attr.attrib(type=Any)
```
...occurs _before_ the type is given to `SignatureField`, making
`SignatureModel.fields` unreliable as a primary source of the form of a handler/
dependency function parameter or return annotation.

Signature modelling would instead use this datastructure as it's input instead of
parsing the handler/dependency functions again itself.