Implements AbstractDTOInterface.
Defines the interface required by the app for a DTO and the application interactions
with it.

If a `dto` param is resolved for any route handler, it is used to parse the
connection data into a data object to be injected into the handler.

If a `dto` param is resolved for any route handler, and no `return_dto` param is
resolved, the DTO type resolved for `dto` is used for conversion of the handler's
return data to an encodable type.

If `return_dto` is resolved for any handler, it is always used for the return
encoding, irrespective of whether a `dto` param is declared or not.