Fix incorrect handler type annotations.
`HTTPRouteHandler.{resolve_after_response(),resolve_after_request()}` were
annotated to return their `__init__()` param type, failing to recognise that they
are wrapped with `AsyncCallable` before being assigned to the handler instance.

In addition to correcting the return types of those methods, this PR adds the
ability to optionally generate a `ParsedSignature` for the `AsyncCallable` type,
and does so for `before_request` handlers. The return type annotation handling is
then refactored around this.

TODO: what happens if a `before_request` handler has an optional return type?

Original PR message:

When using a before request handler with a route with  a `Template` response, the return annotation is set to `T` which causes the the response to be serialized as JSON instead .

It seems to stem from this part of the `HTTPRouteHandler` class:

```python
if before_request_handler := self.resolve_before_request():
        before_request_handler_signature = Signature.from_callable(before_request_handler)
        if (
            before_request_handler_signature.return_annotation
            and before_request_handler_signature.return_annotation is not Signature.empty
        ):
            return_annotation = before_request_handler_signature.return_annotation
```