Add the `UnsupportedType` exception.
To be raised by `DTOInterface` implementation if they are associated with a handler/
annotation that they cannot handle.

It is expected that DTO types will make this check on registration, we can codify
this if necessary.

In the future, whether this raises, or falls back to default ser/de handling can be
configurable.