Single DTO type supports multiple handler/annotations combinations.
This makes a single `AbstractDTOFactory` type much more versatile, making it
able to handle both scalar and collection annotations of the narrowed type.

Factory `on_registration()` callback receives the handler, and the annotation
it should represent and builds a dto backend for each annotation and config
that it encounters during route registration.

We also simplify the logic by being more restrictive, throwing errors for
handlers annotated with `AbstractDTOFactory` subtypes, non-homogeneous
collection type annotations, and being narrowed with union types. All things
that we may choose to support, but best to wait for someone to come knocking
with a concrete use-case.