Refactor `DTOInterface`.
This PR includes refactorings to better integrate DTO encoding with existing
serialization patterns, and modifications to the interface to better support
integration with websocket handlers.

# DTOInterface constructor methods

`from_connection()` reverted back to `from_bytes()`, and receives both the
already awaited raw data from the payload, and the connection instance.

Now that `from_bytes()` receives the raw payload data, already extracted from
the connection, it no longer needs to be an async method, improving symmetry
with the `from_data()` constructor method.

`from_data()` now also receives the connection instance, which further
improves symmetry between the two constructor methods.

# Data extraction methods

The `to_encodable_type()` method no longer receives the connection (this isn't
required any more as the constructor methods receive it, and so it can be
stored on the instance to be accessed in this method if it is required by an
implementation).

# serialization.py

An encoding hook has been added for serialization of DTO instances, so DTO
instances can be processed through the usual channel. This wraps any raw
bytes returned from the DTO in `msgspec.Raw`, which nicely resolves an earlier
issue encountered when trying to build this pattern.

These changes are based on discussion in discord development channel on April
14.