Alt SerializationPluginProtocol leveraging the DTO factory.
If annotated type is supported by the plugin, and no DTO is otherwise defined,
the plugin creates an `AbstractDTOFactory` type for the model type and
assigns it to the handler.

Only one DTO type is generated per model type, which is cached by the plugin
and used anytime that model is encountered by the plugin.

This is intended to replace `SerializationPluginProtocol`, however, I've
implemented it using the name `DTOSerializationPluginProtocol` and left all
the other plugin implementation handling in place in order to ease initial
reviews of the implementation.