Decouple `AbstractDTOFactory` from `AbstractDTOBackend`.
There is no need to have a required backend type for each factory type. This
allows us to pick the backend type that best suits our needs per handler,
particularly depending on the media type supported by the handler, and whether
this data can be parsed into the DTO directly from bytes, or not.