Remove support for declaring DTOs as handler types.
This is something that can always be added back in if it turns out to be
wanted, however, the original case for this was when DTO types were pydantic
models. Now that they can be declared on the layers, I don't think it should
be supported until we find a concrete use-case for it.

Additionally, if a user did want to interact with the DTO type directly, they
could use DI and inject it.

Closes #1525.