Refactor `DTOInterface` - a better pattern for websockets.
Based on discussion in #1518.

This pattern emphasises and supports that the lifespan of a DTO instance
should be bound to the lifespan of the connection.

Any connection-based logic can be performed once, and the DTO used to parse
and serialize as often as necessary for the life of the connection.