Moves model parsing into `AbstractDTOBackend`. (#1597)
This PR moves the bulk of the logic for parsing a model type from `AbstractDTOFactory` to
`AbstractDTOBackend`. This is purely a refactor - no functionality change - just moves
functions (that were private anyway) to within the `_backend` module.

The immediate motivation for this is to move the implementation of `renaming_strategy` from the
frontend to the backend. Our configuration for this closely mirrors that of msgspec, and so by
making the field parsing part of the backend, this will allow us to defer field renaming to
msgspec, or use our own implementation of the renaming strategy for other backend types.