Removes `NestedFieldDefinition` datastructure.
This commit adds datastructures for recursive representation of model attribute type annotations
including:

- SimpleType
- UnionType
- CollectionType
- TupleType
- MappingType

`SimpleType` represents a type that is not composed of other types. This type might be a nested
model and has the `SimpleType.transfer_model` optional attribute that holds the generated
field definitions for the nested model and the model type for data transfer.

The remaining types are composed of other types that might be simple, or not. We keep a reference
on those "composite" types that records whether it holds a nested type, or not. The "inner" types
of a composed type are recursively resolved down to `SimpleType`s and any simple type is checked
to determine if it represents a nested model. This means that the dto frontends don't need to
include logic to determine if their fields are nested collections or unions or whatever - the
only time that `AbstractDTOFactory.detect_nested_field()` is called is when we know that it is a
simple type.