DTO: Refactor transfer model production (#1632)
* Adds `NestedFieldDefinition.transfer_model` attribute.

Moves the production of transfer models to represent nested types into the model parsing
procedure.

Having both the nested type and the transfer type available to us when iterating over field
definitions means we don't need to rely on extracting the nested transfer type from the
full transfer model when creating instances.

* Refactor and centralize model data transfer functions.

Refactors so that functions for transfer in either direction use the field definition
data structures as their source of truth. This means we can do away with keeping a reverse
mapping of renamed fields, as we always have access to both the field name and serialization
name.

* Adds `model_fqdn` attrib and `unique_name` property to `FieldDefinition`.

* Removes `NestedFieldDefinition` datastructure.

This commit adds datastructures for recursive representation of model attribute type annotations
including:

- SimpleType
- UnionType
- CollectionType
- TupleType
- MappingType

`SimpleType` represents a type that is not composed of other types. This type might be a nested
model and has the `SimpleType.transfer_model` optional attribute that holds the generated
field definitions for the nested model and the model type for data transfer.

The remaining types are composed of other types that might be simple, or not. We keep a reference
on those "composite" types that records whether it holds a nested type, or not. The "inner" types
of a composed type are recursively resolved down to `SimpleType`s and any simple type is checked
to determine if it represents a nested model. This means that the dto frontends don't need to
include logic to determine if their fields are nested collections or unions or whatever - the
only time that `AbstractDTOFactory.detect_nested_field()` is called is when we know that it is a
simple type.

* Consolidate data transfer functions.

Creates a single function from `_build_data_from_transfer_data()` and
`_build_transfer_instance_from_model`. These two functions were doing virtually the same thing,
just with different source and target model types.

The new function chooses the correct source and destination attribute names, and destination
model types based on the direction of the data transfer.

* Remove unused function.

* Address review issues.

* Adds fixtures setup for backend abc tests.

Creates fixtures for testing so we don't have to test through the dataclass implementation for
abc behavior tests.

Adds test implementation for UnionType transfer type creation.

* Adds TupleType tests.

* Adds MappingType tests.

* Refactors `transfer_instance_data()`.

Factors out logic of transferring data for individual types.

Adds function per `TransferType` subtype to handle logic of transferring data from that type.

* Recursive data transfer of collection types.

* Explicit non-support of unions of nested composite types.

* Fix docs build.

* Add tests for collection transfer type creation.

* Tests for transfer model type annotation conversions.