Enhancement: DTO Factory computed fields.
Only applies to DTO for inbound data, not return dto.

Compute the value for a model attribute with a callable with an arbitrary number of args.

The transfer model is produced with attributes that represent the args of the callable, not the model attribute.

Values are validated by the transfer model and passed to the callable to retrieve a value for instantiation of the model.

We validate that the return annotation of the callable is consistent with the model attribute that it is populating (TODO).