`ParsedType.is_subclass_of(X)` true for union if all inner types are subtypes of X.
When `ParsedType` was introduced, any union was deliberately left to return `False` with
the intention of waiting for some use-cases to arrive.

This PR closes #1652 where a handler may be typed to return a union of multiple response
types. If all response types are `Response` subtypes then the correct response handler
will be applied.

If any of the union types are not response subtypes then behavior is still incorrect,
but we should either handle that with validation or a special union response handler.