Fix: `simple_asdict()` with attribute value having dataclass type.
`simple_asdict()` was using `isinstance(value, DataclassProtocol)` to determine
if a value was a nested dataclass instance, however this would also return
`True` for dataclass types, which would cause an issue when the dataclass being
converted to a dict had an attribute value that was a dataclass type.

We add a utility function for specifically checking for dataclass instances,
`is_dataclass_instance()`. The logic of this function is pinched from the
private `_is_dataclass_instance()` function inside the stdlib `dataclasses`
module.