Refactor backend TransferFieldDefinition handling.
This commit modifies the way that DTO backends manage their `TransferFieldDefinition` instances.

Previously, the backend would not keep a field definition for excluded fields, however, this proved
limiting when wanting to use the `factory._backends.utils.transfer_instance_data()` method to move
data from unstructured builtins to data model instances. In this scenario, and generally whenever we
are moving inbound data from a transfer model to the data model, we need to transfer a value for all
fields, including fields that are excluded from request and response data, unless an omitted field has
a default value.

Adds an `is_excluded` attribute to `TransferFieldDefinition` so that generated transfer models can
still omit the field, and so that the field can still be omitted from return data.

We no longer modify the `TransferFieldDefinition` type and defaults in the case of `is_partial`.
Instead, the backend implementations do the `UnsetType` splicing at transfer model generation time.
This is important so that we can still discern when a field definition originally had a default value
specified, after the partial handling has been applied.