SQLAlchemyDTO: Improved handling of `hybrid_property`.
In addition to the read-only `FieldDefinition` generated for the getter
method of the hybrid property, if a setter method is defined an additional
write-only `FieldDefinition` is generated that the typed to the argument of
the setter callable.

Fixes issue where expression method names were being added to the DTO fields
as field types (closes #1769).

Adds the `Mark.WRITE_ONLY` field mark for write-only fields.