DTO support for arbitrary generic wrapper types.
Where a handler returns a type that is not supported by the DTO, but:

- the return type is generic
- it has a generic type argument that is supported by the dto
- the type argument maps to an attribute on the return type

...we perform the DTO operations on the data retrieved from that attribute
of the instance returned from the handler, and return the instance.

The constraints are:
- that the type returned from the handler must be a type that litestar can
natively encode
- the annotation must be a type that DTOs can otherwise manage
- the generic type argument of the wrapper class that represents the DTO
managed data should be a complete annotation that fully describes the data.

An example of the last constraint is that the wrapper type should be:

```py
class Wrapper(Generic[T]):
    data: T

@get(dto=...)
def return_wrapped_collection() -> Wrapper[list[Data]]:
    ...

@get(dto=...)
def return_wrapped_scalar() -> Wrapper[Data]:
    ...
```
and not
```py
class CollectionWrapper(Generic[T]):
    data: List[T]

@get(dto=...)
def return_wrapped_collection() -> WrappedCollection[Data]:
    ...
```

This PR handles our pagination types, however, they do violate that last
constraint, where the generic type argument isn't the complete type of the
dto managed data. So for this specific case, we wrap the generic type in
`List`.

Closes #1798