Refactor: moves `Response.get_serializer()` to `serialization.get_serializer()`. (#1820)
This is prep for #1790, I've split the change out for clearer discussion.

The PR removes the `Response.get_serializer()` method in favor of a
`get_serializer()` function in `serialization.py`.

In the current `Response` implementation, `get_serializer()` is called on the
response inside `Response.__init__()`, and the merging of class-level
`type_encoders` with `Response` argument `type_encoders` is done inside the
`get_serializer()` method.

In the #1790 version of `Response`, the response body is not encoded until after
the response object has been returned from the handler, and it is converted into
a low-level `ASGIResponse` object. Due to this, there is still opportunity for
the handler layer resolved `type_encoders` object to be merged with the
`Response` defined `type_encoders` - hence I've removed the logic to merge the
`Response` class and arg defined type encoders from within the
`Response.get_serializer()` method in that PR.

By splitting out the `get_serializer()` method from the response class, this
also seemed to clean up a bit of logic for the logging middleware, where we can
non-optionally return a `Serializer` instance from the
`get_serializer_from_scope()` method.

Finally, having the function in the `serialization.py` module improves
encapsulation of the serialization logic, evidenced by the fact that it removes
import of `DEFAULT_TYPE_ENCODERS` from the module in one instance, and
`default_serializer` in two instances.

So, long story short, I felt this was a decent enough tidy up to warrant a PR
independent of #1790.