Refactor: openapi handler media types
This PR introduces two behavioural differences:

1. template response media types aren't necessarily always HTML.
2. don't mutate the state of the handler media type configuration during schema
generation.

Regarding the first change. If the return annotation of the handler is a `Template`
response, currently we use `MediaType.HTML` for the response content, ignoring what
may have already been defined on the handler. This PR only defaults to
`MediaType.HTML` if it isn't already defined. I believe this is correct, as
template documents don't have to be HTML, although that would be the most common.

Regards the 2nd change, instead of mutating the handler inside the openapi
generation, this PR introduces a local variable to carry the media type for the
function. This is not fixing a known issue, I just noticed the behavior while
working on an earlier PR and think that unnecessarily mutating the configuration
of the handler might be a source of confusion.