Refactor: remove set-cookie ops on genric asgi responses.
Current behavior is to set cookies on asgi responses from handlers if the response
type has a `set_cookie()` method. This looks like it is specific handling for
starlette responses, as they are the only tests that broke on its removal.

If we want to continue to explicitly support this for starlette response types, we
can create a specific response handler for them that does set cookies, however, IMO
we have diverged far enough from starlette to not warrant the specific logic.