refactor: Centralize config of pyright.
All exclusions and other configs should be set in pyproject.toml.

This PR makes an effort to minimize the amount of excluded modules. So, in addition to moving the exclude config from .pre-commit-config.yaml to pyproject.toml, it also omits some of the exclusions that were in .pre-commit-config.yaml and addresses the arising type issues.