fix: Prevent `ASGIConnection.session` from returning `Empty`.
Property is typed to return `dict[str, Any]` but we use `Empty` as a sentinel to instruct session middleware to clear the session.

This PR changes the behavior of `ASGIConnection.session` to raise an exception on session access, just as if the session key is not set at all.

In addition, includes some related refactorings:

- Defines a new constant, `SCOPE_SESSION_KEY: Final = "session"` and replaces uses of `"session"` literal with that.
- Refactors tests/unit/test_connection/test_base.py to utilise fixtures.

Closes #2435