refactor: centralize config of mypy (#2457)
* Remove mypy excludes.

* refactor: centralize config of mypy.

We should favor setting all exclusions and other configs in pyproject.toml. As such, this PR removes the exclude hook config for the pre-commit pyright. Ideally, there should be no difference between running `pre-commit run mypy --all-files` and `pdm run mypy`, and I propose that centralizing the config in this manner is the best way to achieve that.

This PR makes an effort to minimize the amount of excluded modules. So, in addition to moving the exclude config from .pre-commit-config.yaml to pyproject.toml, it also omits some of the exclusions and addresses the arising type issues.

* Address review items.

* Add pre-commit files/excludes and audit additional_dependencies.

* Fix spurious warning.

Pre-commit runs mypy runs with `--ignore-missing-imports` by default, causing an `unused-ignore` warning, where one should exist.