refactor: remove use of is_class_and_subclass() from FieldDefinition.is_subclass_of()
Using `is_class_and_subclass()` is dangerous, because it hides things. Simply returning `False` if a type isn't a class isn't necessarily what we want.

A good example of this is that one of the things that we needed to fix for this PR was to add `ClassVar` to our set of wrapper types. Just like `Required` et al., `ClassVar` isn't an alias for a builtin, it is a type that provides more information to type checkers. However, due to `is_class_and_subclass()` `FieldDefinition.is_subclass_of(int)` would return `True` for `Required[int]`, but `False` for `ClassVar[int]`. This is something that we would have caught earlier were it not for `is_class_and_subclass()`.

The other change required to get this through was to return `False` from `FieldDefinition.is_subclass_of()` if the type is a `Literal`. This is not correct, but is the same as current behavior, and I want to take some time to think about how we should handle these.