fix: handle non UTF-8 data when logging requests (#2530)
When structlog is not installed, the body will not get parsed and
shown as a byte sequence. This byte sequence is then serialized
into a string with the assumption that it is valid UTF-8. This is
not always the case, eg. when uploading a file.

Therefore, use backslashreplace to handle invalid UTF-8 sequences.

Fixes #2529

Co-authored-by: Peter Schutt <peter.github@proton.me>