refactor: use `{get,set}_litestar_scope_state()` for storing csrf tokens.
Current behavior is to add the token directly into scope with a private key, "_csrf_token". However, this causes type errors as scope is a typed dict and does not define that key.

We have an internal namespace inside scope state, this PR refactors csrf token management to use it.

Closes #2499