build: implement covdefaults coverage plugin. (#2528)
This PR adds the [covdefaults](https://github.com/asottile/covdefaults) plugin to our coverage config.

The plugin offers default configs for `[tool.coverage.run]` and `[tool.coverage.report]`.

For "run", the main difference is that it activates branch coverage. This is something I always run on other projects and I think would be a nice addition to our config. Our overall coverage score takes a hit, but it's not as bad as I thought it might be.

For reporting, it adds a `fail_under` value - which we don't currently have as we let sonar handle that. We can override their default of 100, so I set it to 96 which is our current baseline after branch coverage activated.