fix: FieldDefinition.from_kwarg() with None default
This PR fixes an issue where parameters with `None` defaults would have `FieldDefinition.default == Empty`.

Fixing this only broke 2 tests, which were both changed to test that the field def. has an empty default when they clearly shouldn't (i.e., default defined in the test func signature for the param) :shrug:

This PR changes the logic in `FieldDefinition.from_kwarg()` to not filter out the default value from params it passes to `FieldDefinition.from_annotation` when its value is `None`.