fix: openapi schema for literal and enum unions with None.
Existing behavior is to make the schema for every type that is a union with `None` a "one_of" schema, that includes `OpenAPIType.NULL` in the "one_of" types.

https://github.com/litestar-org/litestar/blob/5db8d81fbf24788d38d210e90300e5f0926f830a/litestar/_openapi/schema_generation/schema.py#L319-L340

When a `Literal` or `Enum` type is in a union with `None`, this behavior is not desirable, as we want to have `null` available in the list of available options on the type's schema.

This PR changes literal and enum schema generation so that it can be identified that the types are in a union with `None`, allowing `null` to be included in `Schema.enum` values.

Closes #2456