refactor: remove _openapi.schema_generation.utils.sort_schemas_and_refrences() (#2570)
* refactor: remove _openapi.schema_generation.utils.sort_schemas_and_references()

This method has been at the root of a couple of issues recently:
- #2460
- #2563

We pass the `Schema` instances that would go into a `oneOf` schema property into this function, which attempts to sort the contents by the `Schema.type` property. However, not all `Schema` instances define that property, and we get errors due to trying to do a value comparison against `None`. E.g., where the values of a schema can be other `oneOf` schemas.

Unfortunately, git blame for the commit that adds `sort_schemas_and_references()` points to a 61 file commit that adds attrs support to `Partial`, so hard to draw conclusions as to why it was added from that.

The reason I've opted to remove, is that it is unclear how memebers of the schema should be sorted when they don't have a type, and it is not clear to me whether they even should be sorted, e.g., if a handler returns a `tuple[str, int]`, should those types be sorted alphabetically.

So, this PR removes use of the utility, and modifies the tests that were relying on the schema ordering it provided.

Closes #2563

* Fix flaky test by not relying on deterministic order of union type params.

* refactor: remove _openapi.schema_generation.utils.sort_schemas_and_refrences() (Sourcery refactored) (#2579)

'Refactored by Sourcery'

Co-authored-by: Sourcery AI <>

---------

Co-authored-by: sourcery-ai[bot] <58596630+sourcery-ai[bot]@users.noreply.github.com>
Co-authored-by: Cody Fincher <204685+cofin@users.noreply.github.com>